+++ src/PoshEmail.psm1
@@ -538,7 +538,12 @@ function Invoke-CommandWithEmailWrapper {
         [parameter(ParameterSetName="Script",Mandatory=$false)]
         [parameter(ParameterSetName="ScriptBlock",Mandatory=$false)]
         # Indicates that the cmdlet uses the Secure Sockets Layer (SSL) protocol to establish a connection to the remote computer to send mail. Defaults to $true
-        [bool]$EmailUseSsl = $true
+        [bool]$EmailUseSsl = $true,
+
+        [parameter(ParameterSetName="Script",Mandatory=$false)]
+        [parameter(ParameterSetName="ScriptBlock",Mandatory=$false)]
+        # Some non-PowerShell commands send non-error output to PowerShell's error or warning streams. Adding this option will redirect all streams to output to prevent this.
+        [bool]$RedirectStreams = $false
     )
 
     process {
@@ -576,10 +581,18 @@ function Invoke-CommandWithEmailWrapper {
                 $TempFilePath = "$env:TMPDIR/EmailWrappedCommand.ps1"
             }
             Out-File -FilePath $TempFilePath -InputObject $ScriptBlock
-            $InvokeCommandParams += @{ ScriptBlock = { & $TempFilePath *>&1 } }
+            if ($RedirectStreams) {
+                $InvokeCommandParams += @{ ScriptBlock = { & $TempFilePath *>&1 } }
+            } else {
+                $InvokeCommandParams += @{ ScriptBlock = { & $TempFilePath } }
+            }
         }
         if ($Script) {
-            $InvokeCommandParams += @{ ScriptBlock = { & $Script *>&1 } }
+            if ($RedirectStreams) {
+                $InvokeCommandParams += @{ ScriptBlock = { & $Script *>&1 } }
+            } else {
+                $InvokeCommandParams += @{ ScriptBlock = { & $Script } }
+            }
         }
         if ($ComputerName) {
             $InvokeCommandParams += @{ ComputerName = $ComputerName }